requirements.txt
matplotlib>=3.1.2
numpy>=1.17.4
PyOpenGL>=3.1.4
PyOpenGL-accelerate>=3.1.4
PyQt5>=5.13.2
PyQt5-sip>=12.7.0
scipy>=1.3.3
tripy
quadpy>=0.13.6
Shapely>=1.7.0